feat: implement supplier entity and product unit type

+++ src/main/java/com/eduribeiro8/LilMarket/dto/BatchRequestDTO.java
@@ -19,9 +19,9 @@ public record BatchRequestDTO(
         @Schema(description = "Data de validade", example = "2026-06-01", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotNull LocalDate expirationDate,
         @Schema(description = "Quantidade inicial em estoque", example = "100", requiredMode = Schema.RequiredMode.REQUIRED)
-        @NotNull @PositiveOrZero Integer quantityInStock,
+        @NotNull @PositiveOrZero BigDecimal quantityInStock,
         @Schema(description = "Quantidade perdida", example = "0")
-        @PositiveOrZero Integer quantityLost,
+        @PositiveOrZero BigDecimal quantityLost,
         @Schema(description = "Preço de compra por unidade", example = "1.50", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotNull @Positive BigDecimal purchasePrice
         ) {
